+++ packages/@expo/cli/src/start/server/middleware/__tests__/ManifestMiddleware-test.ts
@@ -1,4 +1,5 @@
 import { getConfig } from '@expo/config';
+import { vol } from 'memfs';
 
 import { asMock } from '../../../../__tests__/asMock';
 import * as Log from '../../../../log';
@@ -93,7 +94,7 @@ describe('checkBrowserRequestAsync', () => {
         // NOTE(EvanBacon): Browsers won't pass the `expo-platform` header so we need to
         // provide the `platform=web` query parameter in order for the multi-platform dev server
         // to return the correct bundle.
-        '/index.bundle?platform=web&dev=true&hot=false&lazy=true',
+        '/index.bundle?platform=web&dev=true&hot=false',
       ],
     });
     expect(res.setHeader).toBeCalledWith('Content-Type', 'text/html');
@@ -127,6 +128,14 @@ describe('checkBrowserRequestAsync', () => {
 });
 
 describe('_getBundleUrl', () => {
+  beforeEach(() => {
+    vol.reset();
+  });
+
+  afterAll(() => {
+    vol.reset();
+  });
+
   const createConstructUrl = () =>
     jest.fn(({ scheme, hostname }) => `${scheme}://${hostname ?? 'localhost'}:8080`);
   it('returns the bundle url with the hostname', () => {
@@ -138,13 +147,36 @@ describe('_getBundleUrl', () => {
         mainModuleName: 'index',
         platform: 'android',
       })
-    ).toEqual(
-      'http://evanbacon.dev:8080/index.bundle?platform=android&dev=true&hot=false&lazy=true'
-    );
+    ).toEqual('http://evanbacon.dev:8080/index.bundle?platform=android&dev=true&hot=false');
 
     expect(constructUrl).toHaveBeenCalledWith({ hostname: 'evanbacon.dev', scheme: 'http' });
   });
   it('returns the bundle url in production', () => {
+    const constructUrl = createConstructUrl();
+    const middleware = new MockManifestMiddleware('/', {
+      constructUrl,
+      mode: 'production',
+      minify: true,
+    });
+    expect(
+      middleware._getBundleUrl({
+        mainModuleName: 'node_modules/expo/AppEntry',
+        platform: 'ios',
+      })
+    ).toEqual(
+      'http://localhost:8080/node_modules/expo/AppEntry.bundle?platform=ios&dev=false&hot=false&minify=true'
+    );
+
+    expect(constructUrl).toHaveBeenCalledWith({ hostname: undefined, scheme: 'http' });
+  });
+
+  it('returns the bundle url in production with lazy enabled', () => {
+    vol.fromJSON(
+      {
+        'node_modules/@expo/metro-runtime/package.json': '',
+      },
+      '/'
+    );
     const constructUrl = createConstructUrl();
     const middleware = new MockManifestMiddleware('/', {
       constructUrl,